refactor(app): make app directory
