+++ src/app/features/app/app.module.ts
@@ -1,15 +1,14 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { HttpClientModule } from '@angular/common/http';
 import { NgModule } from '@angular/core';
-import { ReactiveFormsModule } from '@angular/forms';
 
 import { AppComponent } from './app.component';
 import {
   AppRoutingModule,
   routedComponents as mainAppRoutedComponents,
 } from './app-routing.module';
-import { CoreModule } from './infrastructure/core/core.module';
-import { PipeModule } from './infrastructure/core/pipes/pipe.module';
+import { CoreModule } from '../../infrastructure/core/core.module';
+import { SharedModule } from '../../infrastructure/shared/shared.module';
 
 @NgModule({
   declarations: [AppComponent, mainAppRoutedComponents],
@@ -17,8 +16,7 @@ import { PipeModule } from './infrastructure/core/pipes/pipe.module';
     BrowserModule,
     CoreModule,
     HttpClientModule,
-    PipeModule.forRoot(),
-    ReactiveFormsModule,
+    SharedModule.forRoot(),
     AppRoutingModule,
   ],
   providers: [],
